refactor: (#56) Main ranking 기능 UI 구조 분리
