+++ src/pages/main/model/types.ts
@@ -1,15 +1,5 @@
 export type MainTab = 'ROOM' | 'LUNCH' | 'RANKING' | 'BOARD';
 
-export interface MainRankingItem {
-  id: number;
-  rank: number;
-  title: string;
-  cafeteriaName: string;
-  mealTime: string;
-  likedCount: number;
-  changeText: string;
-}
-
 export interface MainBoardPost {
   id: number;
   category: 'FREE' | 'REVIEW' | 'INFO' | 'TALK';
