+++ src/pages/main/ui/layout/MainTabSection.tsx
@@ -2,7 +2,7 @@ import { Plus } from 'lucide-react';
 import type { MainTab } from '../../model/types';
 import MainBoardSection from '../MainBoardSection';
 import MainLunchMenuSection from '../lunch/MainLunchMenuSection';
-import MainRankingSection from '../MainRankingSection';
+import MainRankingSection from '../ranking/MainRankingSection';
 import { mockRooms } from '../room/mockRooms';
 import RoomCard from '../room/RoomCard';
 import RoomSummary from '../room/RoomSummary';
