+++ src/pages/main/ui/ranking/types.ts
@@ -1,0 +1,9 @@
+export interface MainRankingItem {
+  id: number;
+  rank: number;
+  title: string;
+  cafeteriaName: string;
+  mealTime: string;
+  likedCount: number;
+  changeText: string;
+}
